test: move old utils/ to test_utils/
