+++ backend/app/tests/pytest.ini
@@ -2,3 +2,5 @@
 filterwarnings =
     ignore::pydantic.PydanticDeprecatedSince20:redis_om
     ignore::pydantic.PydanticDeprecatedSince211:redis_om
+
+asyncio_default_fixture_loop_scope = function
